refactor

+++ src/components/modules/traffic/common/SignalSyncUI.vue
@@ -1,8 +1,8 @@
 <template>
   <div>
-    <v-sheet outlined color="white" rounded v-if="show">
+    <v-sheet outlined color="green" rounded v-if="show">
       <v-card>
-        <v-system-bar window dark color="#1A237E" class="mb-2">
+        <v-system-bar window dark class="mb-2">
           <v-icon>mdi-traffic-light</v-icon>
           <span class="overline">Signal Controller</span>
           <v-spacer></v-spacer>
@@ -15,7 +15,7 @@
             <div class="d-flex align-center">
               <MenuButton
                 :items="sync_menu_items"
-                tooltip="Sync ..."
+                tooltip="Sync With ..."
                 icon="mdi-sync"
                 color="teal darken-3"
                 :normal="true"
@@ -29,22 +29,23 @@
             <div>
               <v-btn small color="primary" @click="onFetchAll" :loading="fetchLoading">
                 <v-icon small left>mdi-tray-arrow-down</v-icon>
-                Fetch all
+                Fetch all patterns
               </v-btn>
-              <v-btn small color="warning darken-5" class="ml-3" @click="onRevertBack" :loading="revertLoading">
+
+              <v-btn small color="purple darken-5" class="ml-8 mr-10" @click="onRevertBack" :loading="revertLoading">
                 <v-icon small left>mdi-arrow-u-left-top</v-icon>
                 Revert to Tactics
               </v-btn>
-            </div>
 
-            <v-tooltip top>
-              <template v-slot:activator="{ on }">
-                <v-btn small fab color="primary darken-2" v-on="on" @click="onCreateNewPattern">
-                  <v-icon>mdi-plus</v-icon>
-                </v-btn>
-              </template>
-              <span>Create New Pattern</span>
-            </v-tooltip>
+              <v-tooltip top>
+                <template v-slot:activator="{ on }">
+                  <v-btn small fab class="ml-10" color="primary darken-2" v-on="on" @click="onCreateNewPattern">
+                    <v-icon>mdi-plus</v-icon>
+                  </v-btn>
+                </template>
+                <span>Create New Pattern</span>
+              </v-tooltip>
+            </div>
           </div>
         </v-container>
       </v-card>
@@ -78,9 +79,11 @@ export default {
     fetchLoading: false,
     revertLoading: false,
     lastSyncTime: null,
+    currentPattern: null,
     sync_menu_items: [
       { title: 'Sync with controller', action: 'controller' },
-      { title: 'Sync with Tactics', action: 'tactics' }
+      { title: 'Sync with Tactics', action: 'tactics' },
+      { title: 'Sync with database', action: 'database' }
     ]
   }),
 
@@ -97,9 +100,11 @@ export default {
 
   mounted() {
     this.$bus.$on('APPLY_PATTERN', ({ action, pattern }) => {
-      if (this.controller.ip) {
-        this.actionSignalPattern(action, this.controller.ip, pattern);
-      }
+      this.actionSignalPattern(action, pattern);
+    });
+
+    this.$bus.$on('APPLY_PATTERNS', patterns => {
+      this.onExecutePatterns(patterns);
     });
   },
 
@@ -107,6 +112,18 @@ export default {
     async onSync() {
       if (await this.$refs.confirm.open('Confirm', 'Are you sure you want to sync with the controller?')) {
         this.getCurrentPattern(this.controller.ip);
+      }
+    },
+
+    async onSyncTactics() {
+      if (await this.$refs.confirm.open('Confirm', 'Are you sure you want to sync with the Tactics?')) {
+        //this.getCurrentPattern(this.controller.ip);
+      }
+    },
+
+    async onSyncDatabase() {
+      if (await this.$refs.confirm.open('Confirm', 'Are you sure you want to sync with the database?')) {
+        //this.getCurrentPattern(this.controller.ip);
       }
     },
 
@@ -139,29 +156,42 @@ export default {
           this.onSync();
           break;
         case 'tactics':
-          this.onSync();
+          this.onSyncTactics();
+          break;
+        case 'database':
+          this.onSyncDatabase();
           break;
       }
     },
 
-    async actionSignalPattern(action, ip, signalItem) {
-      if (action === 0) {
-        const message = `Are you sure you want to apply the pattern <strong>${signalItem.pattern}</strong> onto the controller?`;
-        if (this.$refs.confirm && (await this.$refs.confirm.open('Confirm', message))) {
-          const patternNumber = parseInt(signalItem.pattern);
-          this.executePattern(ip, patternNumber);
-        }
-      } else if (action === 1) {
-        const message = `Are you sure you want to modify the pattern <strong>${signalItem.pattern}</strong> and apply onto the controller?`;
-        if (this.$refs.confirm && (await this.$refs.confirm.open('Confirm', message))) {
-          if (await this.$refs.pattern.open('Modify Pattern', signalItem, { constant: true })) {
-            this.changePattern(ip, this.$refs.pattern.get());
+    async onExecutePatterns(patterns) {
+      const message = `Are you sure you want to apply the pattern onto all the controllers listed below?`;
+      if (this.$refs.confirm && (await this.$refs.confirm.open('Confirm', message))) {
+        this.executePatterns(patterns);
+      }
+    },
+
+    async actionSignalPattern(action, signalItem) {
+      const ip = this.controller.ip;
+      if (ip) {
+        if (action === 0) {
+          const message = `Are you sure you want to apply the pattern <strong>${signalItem.pattern}</strong> onto the controller?`;
+          if (this.$refs.confirm && (await this.$refs.confirm.open('Confirm', message))) {
+            const patternNumber = parseInt(signalItem.pattern);
+            this.executePattern(ip, patternNumber);
           }
-        }
-      } else if (action === 2) {
-        const message = `Are you sure you want to delete the pattern <strong>${signalItem.pattern}</strong> from the controller?`;
-        if (this.$refs.confirm && (await this.$refs.confirm.open('Confirm', message))) {
-          this.deletePattern(ip, signalItem);
+        } else if (action === 1) {
+          const message = `Are you sure you want to modify the pattern <strong>${signalItem.pattern}</strong> and apply onto the controller?`;
+          if (this.$refs.confirm && (await this.$refs.confirm.open('Confirm', message))) {
+            if (await this.$refs.pattern.open('Modify Pattern', signalItem, { constant: true })) {
+              this.changePattern(ip, this.$refs.pattern.get());
+            }
+          }
+        } else if (action === 2) {
+          const message = `Are you sure you want to delete the pattern <strong>${signalItem.pattern}</strong> from the controller?`;
+          if (this.$refs.confirm && (await this.$refs.confirm.open('Confirm', message))) {
+            this.deletePattern(ip, signalItem);
+          }
         }
       }
     },
@@ -202,6 +232,16 @@ export default {
       }
     },
 
+    async executePatterns(patterns) {
+      try {
+        const res = await Api.executePatterns({ data: patterns });
+        let data = this.parseResponseData(res);
+        this.$store.dispatch('setSystemStatus', { text: data.Message, color: 'info' });
+      } catch (error) {
+        this.$store.dispatch('setSystemStatus', { text: error, color: 'error' });
+      }
+    },
+
     async executePattern(ip, patternNumber) {
       try {
         const res = await Api.executePattern(ip, patternNumber);
@@ -228,12 +268,16 @@ export default {
       this.syncLoading = true;
       try {
         const res = await Api.getCurrentPatternNumber(ip);
-        const res1 = Api.getCurrentPatternParameters(ip);
-
-        this.parseResponseData(res);
-        // this.parseResponseData(res1);
-
-        this.lastSyncTime = new Date();
+        const data = this.parseResponseData(res);
+        if (data) {
+          if (data.pattern && data.pattern.Code != -1) {
+            this.currentPattern = data.pattern.Message;
+            this.lastSyncTime = new Date();
+            this.$bus.$emit('UPDATE_CURRENT_PATTERN', { pattern: this.currentPattern });
+          } else {
+            this.$store.dispatch('setSystemStatus', { text: data.pattern.Message, color: 'info' });
+          }
+        }
       } catch (error) {
         this.$store.dispatch('setSystemStatus', { text: error, color: 'error' });
       }
@@ -246,7 +290,7 @@ export default {
         this.$bus.$emit('UPDATE_SIGNAL_TABLE', { action: 0, signal: [] });
         const res = await Api.getPatterns(ip, patterns);
         let data = this.parseResponseData(res);
-
+        console.log(data);
         if (data && Object.keys(data.patterns).length > 0) {
           const signals = this.composeSignals(data.patterns);
           this.$bus.$emit('UPDATE_SIGNAL_TABLE', { action: 0, signal: signals });
@@ -270,10 +314,9 @@ export default {
 
     parseResponseData(response) {
       let result = null;
-      console.log(response);
       if (response.status === 200) {
         if (response.data) {
-          console.log(response.data);
+          //console.log(response.data);
           if (response.data.status === 'OK') {
             if (response.data.data !== undefined) {
               result = response.data.data;
@@ -315,7 +358,7 @@ export default {
     },
 
     formatTime(d) {
-      return Utils.formatDateTime(d);
+      return Utils.formatAMPMTime(d);
     }
   }
 };
